requirements.txt
boto3==1.5.6
botocore==1.8.20
Django==2.0
django-ckeditor==5.4.0
django-js-asset==0.1.1
django-storages==1.6.5
docutils==0.14
easy-thumbnails==2.5
jmespath==0.9.3
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
s3transfer==0.1.12
six==1.11.0
